tests: refactor name and PHPUnit config structure
