+++ tests/Feature/Controllers/TeachersController/ControllerCreateTest.php
@@ -5,14 +5,14 @@ namespace Tests\Feature\Controllers\TeachersController;
 use Domains\Accounts\Database\Factories\UserFactory;
 use Domains\Accounts\Enums\UserRolesEnum;
 use Domains\Accounts\Models\User;
-use Domains\Accounts\Tests\Traits\UserRolesProvider;
+use Domains\Accounts\Tests\DataProviders\UserRolesDataProvider;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class ControllerCreateTest extends TestCase
 {
     use RefreshDatabase;
-    use UserRolesProvider;
+    use UserRolesDataProvider;
 
     private User $director;
