+++ tests/Feature/Controllers/TeachersController/ControllerDeleteTest.php
@@ -6,14 +6,14 @@ use Domains\Accounts\Database\Factories\UserFactory;
 use Domains\Accounts\Enums\UserRolesEnum;
 use Domains\Accounts\Models\User;
 use Domains\Accounts\Repositories\UserRepository;
-use Domains\Accounts\Tests\Traits\UserRolesProvider;
+use Domains\Accounts\Tests\DataProviders\UserRolesDataProvider;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
 class ControllerDeleteTest extends TestCase
 {
-    use UserRolesProvider;
+    use UserRolesDataProvider;
     use RefreshDatabase;
     use WithFaker;
