+++ tests/Feature/Controllers/TeachersController/ControllerShowTest.php
@@ -5,13 +5,13 @@ namespace Tests\Feature\Controllers\TeachersController;
 use Domains\Accounts\Database\Factories\UserFactory;
 use Domains\Accounts\Enums\UserRolesEnum;
 use Domains\Accounts\Models\User;
-use Domains\Accounts\Tests\Traits\UserRolesProvider;
+use Domains\Accounts\Tests\DataProviders\UserRolesDataProvider;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class ControllerShowTest extends TestCase
 {
-    use UserRolesProvider;
+    use UserRolesDataProvider;
     use RefreshDatabase;
 
     private User $director;
