+++ tests/Feature/Controllers/TeachersController/ControllerStoreTest.php
@@ -6,7 +6,7 @@ use Domains\Accounts\Database\Factories\UserFactory;
 use Domains\Accounts\Enums\UserRolesEnum;
 use Domains\Accounts\Models\User;
 use Domains\Accounts\Repositories\UserRepository;
-use Domains\Accounts\Tests\Traits\UserRolesProvider;
+use Domains\Accounts\Tests\DataProviders\UserRolesDataProvider;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
@@ -14,7 +14,7 @@ use Tests\TestCase;
 class ControllerStoreTest extends TestCase
 {
     use RefreshDatabase;
-    use UserRolesProvider;
+    use UserRolesDataProvider;
     use WithFaker;
 
     private User $director;
